python/sec-filing-research/pyproject.toml
[project]
name = "sec-filing-research"
version = "0.1.0"
description = "SEC Filing Research: Extract company and filing metadata from SEC EDGAR using Stagehand and Browserbase"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic",
    "python-dotenv",
    "stagehand",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
